chore: Update dependencies in movie recommender system
This commit updates the dependencies in the movie recommender system. The `requirements.txt` file has been modified to include the latest versions of `keras-tuner` and `kaggle`. This update ensures that the project is using the most up-to-date libraries for optimal performance and compatibility.

Note: This commit does not include any functional changes to the code itself.